\docType{package}
\name{h5linker-package}
\alias{h5linker}
\alias{h5linker-package}
\title{h5linker: Minimal Test Package Linking to hdf5lib}
\description{
A simple package that includes minimal C code calling an HDF5 function.
}
\details{
Its purpose is to test if linking against the 'hdf5lib' package works
correctly during R CMD check or install.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmmr/hdf5lib}
  \item Report bugs at \url{https://github.com/cmmr/hdf5lib/issues}
}

}
\author{
\strong{Maintainer}: Daniel P. Smith \email{dansmith01@gmail.com} (\href{https://orcid.org/0000-0002-2479-2044}{ORCID})

Other contributors:
\itemize{
  \item Alkek Center for Metagenomics and Microbiome Research [copyright holder, funder]
}

}
\keyword{internal}
